requirements.txt
django-grappelli
easy-thumbnails
mysql-python